Rename fs utils module
